refactor(connect): add navigation to Connect to prepare for Discovery tab
