+++ packages/hawtio/src/plugins/connect/Connect.tsx
@@ -1,53 +1,66 @@
-import { Connection, connectService } from '@hawtiosrc/plugins/shared/connect-service'
 import {
-  Button,
-  ButtonVariant,
-  DataList,
-  DataListAction,
-  DataListCell,
-  DataListItem,
-  DataListItemCells,
-  DataListItemRow,
-  Dropdown,
-  DropdownItem,
-  DropdownPosition,
-  KebabToggle,
-  Modal,
-  ModalVariant,
+  Nav,
+  NavItem,
+  NavList,
+  PageGroup,
+  PageNavigation,
   PageSection,
   Popover,
   Text,
   TextContent,
   Title,
-  Toolbar,
-  ToolbarContent,
-  ToolbarItem,
 } from '@patternfly/react-core'
-import { OutlinedQuestionCircleIcon, PluggedIcon, PlusIcon, UnpluggedIcon } from '@patternfly/react-icons'
-import React, { useContext, useEffect, useState } from 'react'
-import { Route, Routes } from 'react-router-dom'
+import { OutlinedQuestionCircleIcon } from '@patternfly/react-icons'
+import React from 'react'
+import { NavLink, Navigate, Route, Routes, useLocation } from 'react-router-dom'
 import './Connect.css'
-import { ConnectImportModal } from './ConnectImportModal'
-import { ConnectModal } from './ConnectModal'
-import { DELETE } from './connections'
 import { ConnectContext, useConnections } from './context'
-import { log } from './globals'
+import { Discover } from './discover/Discover'
+import { pluginPath } from './globals'
 import { ConnectLogin } from './login/ConnectLogin'
+import { Remote } from './remote/Remote'
 
 export const Connect: React.FunctionComponent = () => {
   const { connections, dispatch } = useConnections()
+  const { pathname, search } = useLocation()
+
+  const navItems = [
+    { id: 'remote', title: 'Remote', component: Remote },
+    { id: 'discover', title: 'Discover', component: Discover },
+  ]
+
+  const nav = (
+    <Nav aria-label='Connect Nav' variant='tertiary'>
+      <NavList>
+        {navItems.map(({ id, title }) => (
+          <NavItem key={id} isActive={pathname === `${pluginPath}/${id}`}>
+            <NavLink to={{ pathname: id, search }}>{title}</NavLink>
+          </NavItem>
+        ))}
+      </NavList>
+    </Nav>
+  )
+
+  const routes = navItems.map(({ id, component }) => (
+    <Route key={id} path={id} element={React.createElement(component)} />
+  ))
 
   return (
     <ConnectContext.Provider value={{ connections, dispatch }}>
-      <PageSection id='connect-header' variant='light'>
-        <Title id='connect-header-title' headingLevel='h1'>
-          Connect <ConnectHint />
-        </Title>
-      </PageSection>
+      <PageGroup>
+        <PageSection id='connect-header' variant='light'>
+          <Title id='connect-header-title' headingLevel='h1'>
+            Connect <ConnectHint />
+          </Title>
+        </PageSection>
+        <PageNavigation>{nav}</PageNavigation>
+      </PageGroup>
       <PageSection id='connect-main'>
         <Routes>
+          {routes}
+          {/* connect/login should be hidden to nav */}
           <Route key='login' path='login' element={<ConnectLogin />} />
-          <Route index element={<ConnectContent />} />
+          <Route key='root' path='/' element={<Navigate to={navItems[0]?.id ?? ''} />} />
         </Routes>
       </PageSection>
     </ConnectContext.Provider>
@@ -98,199 +111,5 @@ const ConnectHint: React.FunctionComponent = () => {
     >
       <OutlinedQuestionCircleIcon />
     </Popover>
-  )
-}
-
-const ConnectContent: React.FunctionComponent = () => {
-  const { connections } = useContext(ConnectContext)
-  log.debug('Connections:', connections)
-
-  return (
-    <React.Fragment>
-      <ConnectToolbar />
-      <DataList id='connection-list' aria-label='connection list' isCompact>
-        {Object.entries(connections).map(([name, connection]) => (
-          <ConnectionItem key={name} name={name} connection={connection} />
-        ))}
-      </DataList>
-    </React.Fragment>
-  )
-}
-
-const ConnectToolbar: React.FunctionComponent = () => {
-  const { connections } = useContext(ConnectContext)
-  const [isAddOpen, setIsAddOpen] = useState(false)
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false)
-  const [isImportModalOpen, setIsImportModalOpen] = useState(false)
-
-  const handleAddToggle = () => {
-    setIsAddOpen(!isAddOpen)
-  }
-
-  const handleImportModalToggle = () => {
-    setIsImportModalOpen(!isImportModalOpen)
-  }
-
-  const exportConnections = () => {
-    connectService.export(connections)
-  }
-
-  const initialConnection: Connection = {
-    name: '',
-    scheme: 'http',
-    host: '',
-    port: 8080,
-    path: '/hawtio/jolokia',
-  }
-
-  return (
-    <Toolbar id='connect-toolbar'>
-      <ToolbarContent>
-        <ToolbarItem>
-          <Button variant={ButtonVariant.secondary} onClick={handleAddToggle}>
-            <PlusIcon /> Add connection
-          </Button>
-        </ToolbarItem>
-        <ToolbarItem>
-          <Dropdown
-            key='connect-toolbar-dropdown'
-            isPlain
-            isOpen={isDropdownOpen}
-            toggle={<KebabToggle onToggle={() => setIsDropdownOpen(!isDropdownOpen)} />}
-            dropdownItems={[
-              <DropdownItem key='connect-toolbar-dropdown-import' onClick={handleImportModalToggle}>
-                Import connections
-              </DropdownItem>,
-              <DropdownItem key='connect-toolbar-dropdown-export' onClick={exportConnections}>
-                Export connections
-              </DropdownItem>,
-            ]}
-          />
-        </ToolbarItem>
-      </ToolbarContent>
-      <ConnectModal mode='add' isOpen={isAddOpen} onClose={handleAddToggle} input={initialConnection} />
-      <ConnectImportModal isOpen={isImportModalOpen} onClose={handleImportModalToggle} />
-    </Toolbar>
-  )
-}
-
-type ConnectionItemProps = {
-  name: string
-  connection: Connection
-}
-
-const ConnectionItem: React.FunctionComponent<ConnectionItemProps> = props => {
-  const { dispatch } = useContext(ConnectContext)
-  const { name, connection } = props
-  const [reachable, setReachable] = useState(false)
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false)
-  const [isEditOpen, setIsEditOpen] = useState(false)
-  const [isConfirmDeleteOpen, setIsConfirmDeleteOpen] = useState(false)
-
-  useEffect(() => {
-    const check = () => {
-      connectService.checkReachable(connection).then(result => setReachable(result))
-    }
-    check() // initial fire
-    const timer = setInterval(check, 20000)
-    return () => clearInterval(timer)
-  }, [connection])
-
-  const handleDropdownToggle = () => {
-    setIsDropdownOpen(!isDropdownOpen)
-  }
-
-  const handleEditToggle = () => {
-    setIsEditOpen(!isEditOpen)
-  }
-
-  const handleConfirmDeleteToggle = () => {
-    setIsConfirmDeleteOpen(!isConfirmDeleteOpen)
-  }
-
-  const connect = () => {
-    if (!reachable) {
-      return
-    }
-
-    log.debug('Connecting:', connection)
-    connectService.connect(connection)
-  }
-
-  const deleteConnection = () => {
-    dispatch({ type: DELETE, name })
-    handleConfirmDeleteToggle()
-  }
-
-  const ConfirmDeleteModal = () => (
-    <Modal
-      variant={ModalVariant.small}
-      title='Delete Connection'
-      titleIconVariant='danger'
-      isOpen={isConfirmDeleteOpen}
-      onClose={handleConfirmDeleteToggle}
-      actions={[
-        <Button key='delete' variant='danger' onClick={deleteConnection}>
-          Delete
-        </Button>,
-        <Button key='cancel' variant='link' onClick={handleConfirmDeleteToggle}>
-          Cancel
-        </Button>,
-      ]}
-    >
-      You are about to delete the <b>{name}</b> connection.
-    </Modal>
-  )
-
-  return (
-    <DataListItem key={`connection-${name}`} aria-labelledby={`connection ${name}`}>
-      <DataListItemRow>
-        <DataListItemCells
-          dataListCells={[
-            <DataListCell key={`connection-cell-icon-${name}`} isIcon>
-              {reachable ? <PluggedIcon color='green' /> : <UnpluggedIcon color='red' />}
-            </DataListCell>,
-            <DataListCell key={`connection-cell-name-${name}`}>
-              <b>{name}</b>
-            </DataListCell>,
-            <DataListCell key={`connection-cell-url-${name}`} width={3}>
-              {connectService.connectionToUrl(connection)}
-            </DataListCell>,
-          ]}
-        />
-        <DataListAction
-          id={`connection-actions-${name}`}
-          aria-label={`connection actions ${name}`}
-          aria-labelledby={`${name} connection-actions-${name}`}
-        >
-          <Button
-            key={`connection-action-connect-${name}`}
-            variant='primary'
-            onClick={connect}
-            isDisabled={!reachable}
-            isSmall
-          >
-            Connect
-          </Button>
-          <Dropdown
-            key={`connection-action-dropdown-${name}`}
-            isPlain
-            position={DropdownPosition.right}
-            isOpen={isDropdownOpen}
-            toggle={<KebabToggle onToggle={handleDropdownToggle} />}
-            dropdownItems={[
-              <DropdownItem key={`connection-action-edit-${name}`} onClick={handleEditToggle}>
-                Edit
-              </DropdownItem>,
-              <DropdownItem key={`connection-action-delete-${name}`} onClick={handleConfirmDeleteToggle}>
-                Delete
-              </DropdownItem>,
-            ]}
-          />
-          <ConfirmDeleteModal />
-        </DataListAction>
-      </DataListItemRow>
-      <ConnectModal mode='edit' isOpen={isEditOpen} onClose={handleEditToggle} input={connection} />
-    </DataListItem>
   )
 }
