+++ packages/hawtio/src/plugins/connect/discover/Discover.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export const Discover: React.FunctionComponent = () => {
+  return null
+}
