+++ packages/hawtio/src/plugins/connect/globals.ts
@@ -1,5 +1,8 @@
 import { Logger } from '@hawtiosrc/core/logging'
 
+export const pluginId = 'connect'
+export const pluginTitle = 'Connect'
+export const pluginPath = '/connect'
 export const pluginName = 'hawtio-connect'
 export const log = Logger.get(pluginName)
