+++ packages/hawtio/src/plugins/connect/index.ts
@@ -3,6 +3,7 @@ import { helpRegistry } from '@hawtiosrc/help/registry'
 import { preferencesRegistry } from '@hawtiosrc/preferences/registry'
 import { Connect } from './Connect'
 import { ConnectPreferences } from './ConnectPreferences'
+import { pluginId, pluginPath, pluginTitle } from './globals'
 import help from './help.md'
 import { isActive, registerUserHooks } from './init'
 
@@ -11,15 +12,15 @@ const order = 11
 export const connect: HawtioPlugin = () => {
   registerUserHooks()
   hawtio.addPlugin({
-    id: 'connect',
-    title: 'Connect',
-    path: 'connect',
+    id: pluginId,
+    title: pluginTitle,
+    path: pluginPath,
     order,
     component: Connect,
     isActive,
   })
-  helpRegistry.add('connect', 'Connect', help, order)
-  preferencesRegistry.add('connect', 'Connect', ConnectPreferences, order)
+  helpRegistry.add(pluginId, pluginTitle, help, order)
+  preferencesRegistry.add(pluginId, pluginTitle, ConnectPreferences, order)
 }
 
 export * from './connections'
